+++ R/6_utils.R
@@ -605,33 +605,37 @@ get_go_result_sim <-
       dplyr::distinct(name, .keep_all = TRUE) %>%
       dplyr::select(-name)
 
-    # cc_sim_matrix <-
-    #   simplifyEnrichment::GO_similarity(go_id = result$ID[result$ONTOLOGY == "CC"],
-    #                                     ont = "CC",
-    #                                     measure = measure.method) %>%
-    #   as.data.frame() %>%
-    #   tibble::rownames_to_column(var = "name1") %>%
-    #   tidyr::pivot_longer(cols = -name1,
-    #                       names_to = "name2",
-    #                       values_to = "sim") %>%
-    #   dplyr::filter(name1 != name2) %>%
-    #   dplyr::filter(sim > sim.cutoff)
-#
-#     name <- apply(cc_sim_matrix, 1, function(x) {
-#       paste(sort(x[1:2]), collapse = "_")
-#     })
-#
-#     cc_sim_matrix <-
-#       cc_sim_matrix %>%
-#       dplyr::mutate(name = name) %>%
-#       dplyr::arrange(name) %>%
-#       dplyr::distinct(name, .keep_all = TRUE) %>%
-#       dplyr::select(-name)
+    cc_sim_matrix <-
+      GO_similarity_internal(go_id = result$ID[result$ONTOLOGY == "CC"],
+                             ont = "CC",
+                             measure = measure.method,
+                             control.method = control.method)
+    cc_obsolete_terms <- attr(cc_sim_matrix, "obsolete_terms")
+    cc_sim_df <-
+      cc_sim_matrix %>%
+      as.data.frame() %>%
+      tibble::rownames_to_column(var = "name1") %>%
+      tidyr::pivot_longer(cols = -name1,
+                          names_to = "name2",
+                          values_to = "sim") %>%
+      dplyr::filter(name1 != name2) %>%
+      dplyr::filter(sim > sim.cutoff)
+
+    name <- apply(cc_sim_df, 1, function(x) {
+      paste(sort(x[1:2]), collapse = "_")
+    })
+
+    cc_sim_df <-
+      cc_sim_df %>%
+      dplyr::mutate(name = name) %>%
+      dplyr::arrange(name) %>%
+      dplyr::distinct(name, .keep_all = TRUE) %>%
+      dplyr::select(-name)
 
     sim_matrix <-
-      rbind(bp_sim_df, mf_sim_df)
+      rbind(bp_sim_df, mf_sim_df, cc_sim_df)
 
-    attr(sim_matrix, "obsolete_terms") <- c(bp_obsolete_terms, mf_obsolete_terms)
+    attr(sim_matrix, "obsolete_terms") <- c(bp_obsolete_terms, mf_obsolete_terms, cc_obsolete_terms)
 
     sim_matrix
   }